Reduce code duplication in timedep forward/reverse (#2987)
* Move timedep functions into single template functions to reduce code duplication
* Fix timedep reverse search termination and costing summing.
